requirements.txt
fastmcp
google-ads
google-cloud-secret-manager